fix the bugs

+++ src/tendisplus/storage/record.cpp
@@ -662,7 +662,7 @@ std::string RecordValue::encode() const {
     // versionEP
     varint = varintEncode(_versionEP + 1);
     value.insert(value.end(), varint.begin(), varint.end());
-    INVARIANT_D(getRealKeyType(_type) != _type || _versionEP == -1);
+    INVARIANT_D(getRealKeyType(_type) != _type || _versionEP == (uint64_t)-1);
 
     // CAS
     // NOTE(vinchen): cas should initialize -1, not zero.
@@ -727,7 +727,7 @@ Expected<RecordValue> RecordValue::decode(const std::string& value) {
     }
     offset += expt.value().second;
     uint64_t versionEP = expt.value().first - 1;
-    INVARIANT_D(getRealKeyType(typeForMeta) != typeForMeta || versionEP == -1);
+    INVARIANT_D(getRealKeyType(typeForMeta) != typeForMeta || versionEP == (uint64_t)-1);
 
     // CAS
     expt = varintDecodeFwd(valueCstr+offset, value.size() - offset);
